requirements.txt
mediapipe==0.10.32
opencv-python==4.10.0.84
pyautogui==0.9.54
pynput==1.8.1
numpy==1.26.4
matplotlib>=3.8.0
scipy>=1.12.0
Pillow>=10.0.0
